Use ChartContainer in SparkLineChart
