+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Added
 
-- Horizontal overflow style to `<StackedAreaChart />`, which were unintentionally removed
+- Horizontal overflow style to `<StackedAreaChart />`, which were unintentionally removed.
+- Use `<ChartContainer />` in `<SparkLineChart />`.
+
 
 ### Changed
