+++ src/components/SparkLineChart/SparkLineChart.scss
@@ -13,7 +13,3 @@
   border: 0 !important;
   // stylelint-enable declaration-no-important
 }
-
-.Container {
-  @include chart-container;
-}
