+++ src/components/SparkLineChart/SparkLineChart.tsx
@@ -1,16 +1,9 @@
-import React, {useState, useLayoutEffect, useCallback} from 'react';
-import {useDebouncedCallback} from 'use-debounce';
-import {scaleLinear} from 'd3-scale';
+import React from 'react';
 
 import type {DataSeries} from '../../types';
-import {useThemeSeriesColors} from '../../hooks/use-theme-series-colors';
-import {useResizeObserver, useTheme} from '../../hooks';
-import {XMLNS} from '../../constants';
+import {ChartContainer} from '../ChartContainer';
 
-import styles from './SparkLineChart.scss';
-import {Series} from './components';
-
-const SVG_MARGIN = 2;
+import {Chart} from './Chart';
 
 export interface Coordinates {
   x: number;
@@ -34,140 +27,16 @@ export function SparkLineChart({
   offsetRight = 0,
   theme,
 }: SparkLineChartProps) {
-  const {
-    ref: containerRef,
-    setRef: setContainerRef,
-    entry,
-  } = useResizeObserver();
-  const [svgDimensions, setSvgDimensions] = useState({width: 0, height: 0});
-  const selectedTheme = useTheme(theme);
-  const seriesColors = useThemeSeriesColors(data, selectedTheme);
-
-  const [updateMeasurements] = useDebouncedCallback(() => {
-    if (entry == null) return;
-
-    setSvgDimensions({
-      height: entry.contentRect.height,
-      width: entry.contentRect.width,
-    });
-  }, 10);
-
-  const handlePrintMediaQueryChange = useCallback(
-    (event: MediaQueryListEvent) => {
-      if (event.matches && entry != null) {
-        setSvgDimensions({
-          height: entry.contentRect.width,
-          width: entry.contentRect.height,
-        });
-      }
-    },
-    [entry],
-  );
-
-  useLayoutEffect(() => {
-    if (entry == null) return;
-
-    if (containerRef == null) return;
-
-    updateMeasurements();
-
-    const isServer = typeof window === 'undefined';
-
-    if (!isServer) {
-      window.addEventListener('resize', () => updateMeasurements());
-
-      if (typeof window.matchMedia('print').addEventListener === 'function') {
-        window
-          .matchMedia('print')
-          .addEventListener('change', handlePrintMediaQueryChange);
-      } else if (typeof window.matchMedia('print').addListener === 'function') {
-        window.matchMedia('print').addListener(handlePrintMediaQueryChange);
-      }
-    }
-
-    return () => {
-      if (!isServer) {
-        window.removeEventListener('resize', () => updateMeasurements());
-
-        if (typeof window.matchMedia('print').addEventListener === 'function') {
-          window
-            .matchMedia('print')
-            .removeEventListener('change', handlePrintMediaQueryChange);
-        } else if (
-          typeof window.matchMedia('print').addListener === 'function'
-        ) {
-          window
-            .matchMedia('print')
-            .removeListener(handlePrintMediaQueryChange);
-        }
-      }
-    };
-  }, [entry, containerRef, updateMeasurements, handlePrintMediaQueryChange]);
-
-  const {width, height} = svgDimensions;
-
-  const xValues = Array.prototype.concat.apply(
-    [],
-    data.map(({data}) => data.map(({key}) => key)),
-  );
-
-  const minXValues = Math.min(...xValues);
-  const maxXValues = Math.max(...xValues);
-
-  const yValues = Array.prototype.concat.apply(
-    [],
-    data.map(({data}) => data.map(({value}) => value)),
-  );
-
-  const yScale = scaleLinear()
-    .range([height - SVG_MARGIN, SVG_MARGIN])
-    .domain([Math.min(...yValues), Math.max(...yValues)]);
-
   return (
-    <div
-      ref={setContainerRef}
-      className={styles.Container}
-      style={{
-        background: selectedTheme.chartContainer.backgroundColor,
-        padding: selectedTheme.chartContainer.padding,
-        borderRadius: selectedTheme.chartContainer.borderRadius,
-      }}
-    >
-      {accessibilityLabel ? (
-        <span className={styles.VisuallyHidden}>{accessibilityLabel}</span>
-      ) : null}
-
-      <svg xmlns={XMLNS} aria-hidden width={width} height={height}>
-        {data.map((series, index) => {
-          const singleOffsetLeft = series.isComparison ? 0 : offsetLeft;
-          const singleOffsetRight = series.isComparison ? 0 : offsetRight;
-
-          const xScale = scaleLinear()
-            .range([
-              singleOffsetLeft + SVG_MARGIN,
-              width - singleOffsetRight - SVG_MARGIN,
-            ])
-            .domain([minXValues, maxXValues]);
-
-          const seriesWithColor = {
-            color: seriesColors[index],
-            ...series,
-          };
-
-          return (
-            <g key={index}>
-              <Series
-                xScale={xScale}
-                yScale={yScale}
-                data={seriesWithColor}
-                isAnimated={isAnimated}
-                svgDimensions={svgDimensions}
-                theme={selectedTheme}
-              />
-            </g>
-          );
-        })}
-      </svg>
-    </div>
+    <ChartContainer theme={theme}>
+      <Chart
+        data={data}
+        accessibilityLabel={accessibilityLabel}
+        isAnimated={isAnimated}
+        offsetLeft={offsetLeft}
+        offsetRight={offsetRight}
+        theme={theme}
+      />
+    </ChartContainer>
   );
 }
